backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Mzee chakula backend interface"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "chromadb>=1.3.5",
    "email-validator>=2.3.0",
    "fastapi>=0.121.3",
    "hiredis>=3.3.0",
    "httpx>=0.28.1",
    "huggingface-hub>=1.1.5",
    "joblib>=1.5.2",
    "langchain>=1.0.8",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.1",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "prometheus-client>=0.23.1",
    "psutil>=7.1.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pypdf2>=3.0.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sentry-sdk[fastapi]>=2.45.0",
    "sqlalchemy>=2.0.44",
    "supabase>=2.24.0",
    "tavily-python>=0.7.13",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
    "xgboost>=3.1.2",
]
